requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
crewai>=0.28.0
crewai-tools>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
ollama>=0.1.5 